+++ core/src/main/java/org/seedstack/seed/core/internal/lifecycle/LifecycleManager.java
@@ -5,12 +5,17 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
+
 package org.seedstack.seed.core.internal.lifecycle;
+
+import java.lang.reflect.Method;
 
 interface LifecycleManager {
     void started();
 
     void stopping();
+
+    void registerPreDestroy(Object o, Method m);
 
     void registerAutoCloseable(AutoCloseable autoCloseable);